requirements.txt
flask==3.1.0
transformers==4.48.2
datasets==3.2.0
pydantic==2.10.6
langchain==0.3.17
torch==2.6.0
faiss-cpu==1.10.0